scripts/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
simple-salesforce>=1.12.0
slack-sdk>=3.27.0
python-dotenv>=1.0.0
